Add Cup from which users can select a list of Things which are then added and displayed on the respective player's rack.
Add toString for Thing model classes. 

Add Rack object to Player.

Some minor cleanup and refactoring.

Add "Kings & Things" logo to main menu.

Created a custom image view that allows a border to be wrapped around it
using a pane, as well as a custom event handler so that events aren't
consumed by the wrapper pane.